+++ src/test/php/lang/ast/unittest/parse/FunctionsTest.class.php
@@ -5,7 +5,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function empty_function_without_parameters() {
     $this->assertNodes(
-      [['(' => ['a', [], [], [], null, null]]],
+      [['function' => ['a', [[], null], []]]],
       $this->parse('function a() { }')
     );
   }
@@ -13,7 +13,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function two_functions() {
     $this->assertNodes(
-      [['(' => ['a', [], [], [], null, null]], ['(' => ['b', [], [], [], null, null]]],
+      [['function' => ['a', [[], null], []]], ['function' => ['b', [[], null], []]]],
       $this->parse('function a() { } function b() { }')
     );
   }
@@ -21,7 +21,7 @@ class FunctionsTest extends ParseTest {
   #[@test, @values(['param', 'protected'])]
   public function with_parameter($name) {
     $this->assertNodes(
-      [['(' => ['a', [], [[$name, false, null, false, null, null]], [], null, null]]],
+      [['function' => ['a', [[[$name, false, null, false, null, null]], null], []]]],
       $this->parse('function a($'.$name.') { }')
     );
   }
@@ -29,7 +29,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function with_reference_parameter() {
     $this->assertNodes(
-      [['(' => ['a', [], [['param', true, null, false, null, null]], [], null, null]]],
+      [['function' => ['a', [[['param', true, null, false, null, null]], null], []]]],
       $this->parse('function a(&$param) { }')
     );
   }
@@ -37,7 +37,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function dangling_comma_in_parameter_lists() {
     $this->assertNodes(
-      [['(' => ['a', [], [['param', false, null, false, null, null]], [], null, null]]],
+      [['function' => ['a', [[['param', false, null, false, null, null]], null], []]]],
       $this->parse('function a($param, ) { }')
     );
   }
@@ -45,7 +45,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function with_typed_parameter() {
     $this->assertNodes(
-      [['(' => ['a', [], [['param', false, 'string', false, null, null]], [], null, null]]],
+      [['function' => ['a', [[['param', false, 'string', false, null, null]], null], []]]],
       $this->parse('function a(string $param) { }')
     );
   }
@@ -53,7 +53,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function with_nullable_typed_parameter() {
     $this->assertNodes(
-      [['(' => ['a', [], [['param', false, '?string', false, null, null]], [], null, null]]],
+      [['function' => ['a', [[['param', false, '?string', false, null, null]], null], []]]],
       $this->parse('function a(?string $param) { }')
     );
   }
@@ -61,7 +61,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function with_variadic_parameter() {
     $this->assertNodes(
-      [['(' => ['a', [], [['param', false, null, true, null, null]], [], null, null]]],
+      [['function' => ['a', [[['param', false, null, true, null, null]], null], []]]],
       $this->parse('function a(... $param) { }')
     );
   }
@@ -69,7 +69,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function with_optional_parameter() {
     $this->assertNodes(
-      [['(' => ['a', [], [['param', false, null, false, null, ['null' => 'null']]], [], null, null]]],
+      [['function' => ['a', [[['param', false, null, false, null, ['null' => 'null']]], null], []]]],
       $this->parse('function a($param= null) { }')
     );
   }
@@ -77,7 +77,7 @@ class FunctionsTest extends ParseTest {
   #[@test]
   public function with_return_type() {
     $this->assertNodes(
-      [['(' => ['a', [], [], [], 'void', null]]],
+      [['function' => ['a', [[], 'void'], []]]],
       $this->parse('function a(): void { }')
     );
   }
@@ -86,16 +86,16 @@ class FunctionsTest extends ParseTest {
   public function default_closure() {
     $block= ['+' => [['(variable)' => 'a'], ['(literal)' => '1']]];
     $this->assertNodes(
-      [['(' => [null, [], [['a', false, null, false, null, null]], [['return' => $block]], null, null]]],
-      $this->parse('function($a) { return $a + 1; };')
+      [['function' => [[[], null], null, [['return' => $block]]]]],
+      $this->parse('function() { return $a + 1; };')
     );
   }
 
   #[@test]
-  public function default_closure_with_use_by_vaue() {
+  public function default_closure_with_use_by_value() {
     $block= ['+' => [['(variable)' => 'a'], ['(literal)' => '1']]];
     $this->assertNodes(
-      [['(' => [null, [], [], [['return' => $block]], null, ['$a', '$b']]]],
+      [['function' => [[[], null], ['$a', '$b'], [['return' => $block]]]]],
       $this->parse('function() use($a, $b) { return $a + 1; };')
     );
   }
@@ -104,17 +104,8 @@ class FunctionsTest extends ParseTest {
   public function default_closure_with_use_by_reference() {
     $block= ['+' => [['(variable)' => 'a'], ['(literal)' => '1']]];
     $this->assertNodes(
-      [['(' => [null, [], [], [['return' => $block]], null, ['$a', '&$b']]]],
+      [['function' => [[[], null], ['$a', '&$b'], [['return' => $block]]]]],
       $this->parse('function() use($a, &$b) { return $a + 1; };')
-    );
-  }
-
-  #[@test]
-  public function short_closure() {
-    $block= ['+' => [['(variable)' => 'a'], ['(literal)' => '1']]];
-    $this->assertNodes(
-      [['(' => [null, [], [['a', false, null, false, null, null]], [['==>' => $block]], null, null]]],
-      $this->parse('($a) ==> $a + 1;')
     );
   }
 
@@ -122,7 +113,7 @@ class FunctionsTest extends ParseTest {
   public function generator() {
     $statement= ['yield' => [null, null]];
     $this->assertNodes(
-      [['(' => ['a', [], [], [$statement], null, null]]],
+      [['function' => ['a', [[], null], [$statement]]]],
       $this->parse('function a() { yield; }')
     );
   }
@@ -131,7 +122,7 @@ class FunctionsTest extends ParseTest {
   public function generator_with_value() {
     $statement= ['yield' => [null, ['(literal)' => '1']]];
     $this->assertNodes(
-      [['(' => ['a', [], [], [$statement], null, null]]],
+      [['function' => ['a', [[], null], [$statement]]]],
       $this->parse('function a() { yield 1; }')
     );
   }
@@ -140,7 +131,7 @@ class FunctionsTest extends ParseTest {
   public function generator_with_key_and_value() {
     $statement= ['yield' => [['(literal)' => '"number"'], ['(literal)' => '1']]];
     $this->assertNodes(
-      [['(' => ['a', [], [], [$statement], null, null]]],
+      [['function' => ['a', [[], null], [$statement]]]],
       $this->parse('function a() { yield "number" => 1; }')
     );
   }
@@ -149,7 +140,7 @@ class FunctionsTest extends ParseTest {
   public function generator_delegation() {
     $statement= ['yield' => ['[' => []]];
     $this->assertNodes(
-      [['(' => ['a', [], [], [$statement], null, null]]],
+      [['function' => ['a', [[], null], [$statement]]]],
       $this->parse('function a() { yield from []; }')
     );
   }